chore: refactor stages
