+++ crates/amaru/src/stages/ledger.rs
